requirements.txt
asgiref==3.8.1
Brotli==1.1.0
dj-database-url==2.2.0
Django==4.2.16
django-cors-headers==4.4.0
djangocms-installer==2.0.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
whitenoise==6.7.0
